EmotionalSpending/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plaid>=0.1.7",
    "plaid-python>=29.1.0",
    "plotly>=6.0.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.43.2",
]
